'Ground Station Text Panel' Functionality
Added the 'Ground Station Text Panel' functionality as per the software
backlog spreadsheet. This was done by creating 'CR_status" and
'MR_status' global variables that get set in the
'waitForAcknowledgement(' function and are parsed in the
'request_status_Callback(' function. Further work must be done to
actually get the status strings from the serial port but this change is
trivial in MATLAB and will mostly rely on that functionality being
implemented on the MR Arduino