Smart Com Port Detection
Added functionality such that on startup the GUI checks for valid serial
ports between 'COM1' and 'COM40' and puts them in a list for the user to
select. There is some error handling for if a port that was once
available is no longer available when the user switches to it.

Also updated the README with the start of some information about
different commands that can be sent.

Updated GUI close request function so that it makes sure the user didn't
accidentally close it before they intended to.