Image Decoding Python Script Integration
Added the capability to call the 'picDecode.py' script to convert an
image string that was saved to the 'picString.txt' file into an
automatically-named jpg file.

Also changed the directory structure so that support functions are in
their own folder as are image-related files. Another folder can be added
in the future for mission log logs!