GS Mission Log Saves
Added functionality so that if the user selects a checkbox on the GUI
then the mission log will be saved to a text file as it updates. The
text file will be saved to the 'MissionLogs' directory under a filename
constructed from when the mission began.

There is possibly a cleaner way to construct the filename than do string
parsing each time, but this at least works for now and is fast enough
that it doesn't matter too much.

Also added more information to the README file.