Rappelling Test Changes
Added a case statement within waitForAcknowledgement for the rappelling
command. Right now it's basically just for debugging purposes as it will
print out whatever the MR sends it during the rappelling loop.

The timeout for commands was also increased to 2000 seconds for testing
purposes and will have to be reverted in a later commit.