Changes 3/13 From Debugging $SR and $I
Made some changes with how the GS waits for acknowledgements for both
status requests and images.

Found a potential but in the response from an imaging command where the
MR (sometimes) responds with a ' ' character which could be the source
of the seemingly random error that an image cannot be decoded properly.
This will have to be investigated further.