Imaging Testing Changes
Added the capability to parse the received image string for the number
of characters it should contain. The number of characters received
compared to the number of characters that should have been sent is also
displayed to the GS in the command window. This should later be moved to
the mission log.

The GUI still attempts to decode the image even if not all of the
characters were received because sometimes it still works to decode it
if only ~5% or less of the characters were dropped.