Rappelling Testing Changes Part 2
Added a folder to hold the rappelling data that is saved. There is a
ProcessRappelData function that is in progress to parse the data and
display it against our model. This is used for results and discussion
for the AIAA paper.

Also figured out why the "response" during rappelling was displayed
multiple times and removed the bug.