Driving Testing Changes 3/23
Modified the ProcessDriveData algorithm to account for changes in the
output data format.

Also updated the GUI to have a "cancel command" button that only appears
after you have sent a command. THIS DOES NOT ACTUALLY CANCEL THE COMMAND
ON THE CR. It only stops the timeout on the GUI to avoid having to
completely restart the GUI if something is not working correctly on the
command execution.

Also added some checking of the serial port when commands and status
requests are sent to decrease the number of crashes that occur and
streamline usage of the GUI during testing.