Fixed bug(s) in waitForAcknowledgement
It looked like the fprintf for driving data was only doing it when
turning. Swapped that but the functionality of writing to the file is
still commented out. (THE GUI WILL STILL CREATE AN EMPTY FILE THOUGH)

Also found that we were only looking for an $RUP\n for a return command
if it *wasn't* a return command so I swapped that and we should be good
now. It expects to get a $R0P\n for a, auto-rappel, auto-spool or rappel
command.